+++ graphdatascience/session/dedicated_sessions.py
@@ -63,8 +63,7 @@ class DedicatedSessions:
         cloud_location: Optional[CloudLocation] = None,
         timeout: Optional[int] = None,
         neo4j_driver_options: Optional[dict[str, Any]] = None,
-        tls_root_certs: Optional[bytes] = None,
-        disable_server_verification: bool = False,
+        arrow_client_options: Optional[dict[str, Any]] = None,
     ) -> AuraGraphDataScience:
         if db_connection is None:
             if not cloud_location:
@@ -106,8 +105,7 @@ class DedicatedSessions:
             session_bolt_connection_info=session_bolt_connection_info,
             arrow_authentication=arrow_authentication,
             db_runner=db_runner,
-            tls_root_certs=tls_root_certs,
-            disable_server_verification=disable_server_verification,
+            arrow_client_options=arrow_client_options,
         )
 
     def _create_db_runner(
@@ -213,14 +211,12 @@ class DedicatedSessions:
         session_bolt_connection_info: DbmsConnectionInfo,
         arrow_authentication: ArrowAuthentication,
         db_runner: Optional[Neo4jQueryRunner],
-        tls_root_certs: Optional[bytes],
-        disable_server_verification: bool,
+        arrow_client_options: Optional[dict[str, Any]] = None,
     ) -> AuraGraphDataScience:
         return AuraGraphDataScience.create(
             session_bolt_connection_info=session_bolt_connection_info,
             arrow_authentication=arrow_authentication,
             db_endpoint=db_runner,
             delete_fn=lambda: self._aura_api.delete_session(session_id=session_id),
-            arrow_tls_root_certs=tls_root_certs,
-            arrow_disable_server_verification=disable_server_verification,
+            arrow_client_options=arrow_client_options,
         )